Create DocumentSimilarity.py
The task is to compute document similarity and represent semantic information using Vector Space Models. Various semantic features like Term frequency, TF.IDF and Mutual Information. Another approach was to use cluster of similar words. We will also compare binary representations, tf.idf weights and mutual information for determining the coordinates of an object. We will perform a simple evaluation to
compare to what extent the different choices lead to different solutions and to decide which solution is best.