Create LanguageModels.py
The focus of the project was to build a bi-gram language model, then create a language model with a state-of-the-art tool which provides flexible options for dealing with unknown words and smoothing. We then compared the effectiveness of different n-gram and smoothing methods on test data. We measured the effectiveness using Word prediction and Sentence completion tasks.